+++ pkg/ledger/execute_script_test.go
@@ -4,7 +4,10 @@ import (
 	"context"
 	"crypto/sha256"
 	"encoding/json"
+	"fmt"
+	"sort"
 	"strconv"
+	"strings"
 	"testing"
 
 	"github.com/DmitriyVTitov/size"
@@ -21,7 +24,7 @@ func TestNoScript(t *testing.T) {
 	runOnLedger(func(l *ledger.Ledger) {
 		script := core.ScriptData{}
 
-		_, err := l.ExecuteScripts(context.Background(), false, script)
+		_, err := l.ExecuteScript(context.Background(), false, script)
 		assert.IsType(t, &ledger.ScriptError{}, err)
 		assert.Equal(t, ledger.ScriptErrorNoScript, err.(*ledger.ScriptError).Code)
 	})
@@ -33,7 +36,7 @@ func TestCompilationError(t *testing.T) {
 			Script: core.Script{Plain: "willnotcompile"},
 		}
 
-		_, err := l.ExecuteScripts(context.Background(), false, script)
+		_, err := l.ExecuteScript(context.Background(), false, script)
 		assert.IsType(t, &ledger.ScriptError{}, err)
 		assert.Equal(t, ledger.ScriptErrorCompilationFailed, err.(*ledger.ScriptError).Code)
 	})
@@ -51,7 +54,7 @@ func TestSend(t *testing.T) {
 					)`,
 				},
 			}
-			_, err := l.ExecuteScripts(context.Background(), false, script)
+			_, err := l.ExecuteScript(context.Background(), false, script)
 			require.NoError(t, err)
 
 			assertBalance(t, l, "user:001",
@@ -68,7 +71,7 @@ func TestSend(t *testing.T) {
 					)`,
 				},
 			}
-			_, err := l.ExecuteScripts(context.Background(), false, script)
+			_, err := l.ExecuteScript(context.Background(), false, script)
 			require.Error(t, err)
 			require.True(t, ledger.IsValidationError(err))
 			require.ErrorContains(t, err, "transaction has no postings")
@@ -84,7 +87,7 @@ func TestSend(t *testing.T) {
 					)`,
 				},
 			}
-			_, err := l.ExecuteScripts(context.Background(), false, script)
+			_, err := l.ExecuteScript(context.Background(), false, script)
 			require.Error(t, err)
 			require.True(t, ledger.IsValidationError(err))
 			require.ErrorContains(t, err, "transaction has no postings")
@@ -104,10 +107,9 @@ func TestSend(t *testing.T) {
 					)`,
 				},
 			}
-			res, err := l.ExecuteScripts(context.Background(), false, script)
+			res, err := l.ExecuteScript(context.Background(), false, script)
 			require.NoError(t, err)
-			require.Equal(t, 1, len(res))
-			require.Equal(t, 1, len(res[0].Postings))
+			require.Equal(t, 1, len(res.Postings))
 
 			assertBalance(t, l, "user:001",
 				"USD/2", core.NewMonetaryInt(100))
@@ -127,10 +129,9 @@ func TestSend(t *testing.T) {
 					)`,
 				},
 			}
-			res, err := l.ExecuteScripts(context.Background(), false, script)
+			res, err := l.ExecuteScript(context.Background(), false, script)
 			require.NoError(t, err)
-			require.Equal(t, 1, len(res))
-			require.Equal(t, 1, len(res[0].Postings))
+			require.Equal(t, 1, len(res.Postings))
 
 			assertBalance(t, l, "user:001",
 				"USD/2", core.NewMonetaryInt(101))
@@ -155,7 +156,7 @@ func TestNoVariables(t *testing.T) {
 			},
 		}
 
-		_, err := l.ExecuteScripts(context.Background(), false, script)
+		_, err := l.ExecuteScript(context.Background(), false, script)
 		assert.Error(t, err)
 	})
 }
@@ -179,7 +180,7 @@ func TestVariables(t *testing.T) {
 			},
 		}
 
-		_, err := l.ExecuteScripts(context.Background(), false, script)
+		_, err := l.ExecuteScript(context.Background(), false, script)
 		require.NoError(t, err)
 
 		assertBalance(t, l, "user:042",
@@ -198,7 +199,7 @@ func TestVariablesEmptyAccount(t *testing.T) {
 					)`,
 			},
 		}
-		_, err := l.ExecuteScripts(context.Background(), false, script)
+		_, err := l.ExecuteScript(context.Background(), false, script)
 		require.NoError(t, err)
 
 		script = core.ScriptData{
@@ -220,7 +221,7 @@ func TestVariablesEmptyAccount(t *testing.T) {
 				},
 			},
 		}
-		_, err = l.ExecuteScripts(context.Background(), false, script)
+		_, err = l.ExecuteScript(context.Background(), false, script)
 		require.NoError(t, err)
 
 		assertBalance(t, l, "alice", "EUR", core.NewMonetaryInt(1))
@@ -254,7 +255,7 @@ func TestEnoughFunds(t *testing.T) {
 			},
 		}
 
-		_, err = l.ExecuteScripts(context.Background(), false, script)
+		_, err = l.ExecuteScript(context.Background(), false, script)
 		assert.NoError(t, err)
 	})
 }
@@ -285,7 +286,7 @@ func TestNotEnoughFunds(t *testing.T) {
 			},
 		}
 
-		_, err = l.ExecuteScripts(context.Background(), false, script)
+		_, err = l.ExecuteScript(context.Background(), false, script)
 		assert.True(t, ledger.IsScriptErrorWithCode(err, apierrors.ErrInsufficientFund))
 	})
 }
@@ -311,7 +312,7 @@ func TestMissingMetadata(t *testing.T) {
 			},
 		}
 
-		_, err := l.ExecuteScripts(context.Background(), false, script)
+		_, err := l.ExecuteScript(context.Background(), false, script)
 		assert.True(t, ledger.IsScriptErrorWithCode(err, ledger.ScriptErrorCompilationFailed))
 	})
 }
@@ -378,7 +379,7 @@ func TestMetadata(t *testing.T) {
 			},
 		}
 
-		_, err = l.ExecuteScripts(context.Background(), false, script)
+		_, err = l.ExecuteScript(context.Background(), false, script)
 		require.NoError(t, err)
 
 		assertBalance(t, l, "sales:042", "COIN", core.NewMonetaryInt(0))
@@ -451,7 +452,7 @@ func TestSetTxMeta(t *testing.T) {
 	} {
 		t.Run(tc.name, func(t *testing.T) {
 			runOnLedger(func(l *ledger.Ledger) {
-				_, err := l.ExecuteScripts(context.Background(), false, tc.script)
+				_, err := l.ExecuteScript(context.Background(), false, tc.script)
 
 				if tc.expectedErrorCode != "" {
 					require.Error(t, err)
@@ -483,7 +484,7 @@ func TestScriptSetReference(t *testing.T) {
 			Reference: "tx_ref",
 		}
 
-		_, err := l.ExecuteScripts(context.Background(), false, script)
+		_, err := l.ExecuteScript(context.Background(), false, script)
 		require.NoError(t, err)
 
 		last, err := l.GetLedgerStore().GetLastTransaction(context.Background())
@@ -495,7 +496,7 @@ func TestScriptSetReference(t *testing.T) {
 
 func TestScriptReferenceConflict(t *testing.T) {
 	runOnLedger(func(l *ledger.Ledger) {
-		_, err := l.ExecuteScripts(context.Background(), false,
+		_, err := l.ExecuteScript(context.Background(), false,
 			core.ScriptData{
 				Script: core.Script{
 					Plain: `
@@ -509,7 +510,7 @@ func TestScriptReferenceConflict(t *testing.T) {
 			})
 		require.NoError(t, err)
 
-		_, err = l.ExecuteScripts(context.Background(), false,
+		_, err = l.ExecuteScript(context.Background(), false,
 			core.ScriptData{
 				Script: core.Script{
 					Plain: `
@@ -529,7 +530,7 @@ func TestScriptReferenceConflict(t *testing.T) {
 func TestSetAccountMeta(t *testing.T) {
 	runOnLedger(func(l *ledger.Ledger) {
 		t.Run("valid", func(t *testing.T) {
-			res, err := l.ExecuteScripts(context.Background(), false,
+			_, err := l.ExecuteScript(context.Background(), false,
 				core.ScriptData{
 					Script: core.Script{Plain: `
 						send [USD/2 99] (
@@ -544,7 +545,6 @@ func TestSetAccountMeta(t *testing.T) {
   					`},
 				})
 			require.NoError(t, err)
-			require.Equal(t, 1, len(res))
 
 			acc, err := l.GetAccount(context.Background(), "alice")
 			require.NoError(t, err)
@@ -559,7 +559,7 @@ func TestSetAccountMeta(t *testing.T) {
 		})
 
 		t.Run("invalid syntax", func(t *testing.T) {
-			_, err := l.ExecuteScripts(context.Background(), false,
+			_, err := l.ExecuteScript(context.Background(), false,
 				core.ScriptData{
 					Script: core.Script{Plain: `
 						send [USD/2 99] (
@@ -604,7 +604,7 @@ func TestMonetaryVariableBalance(t *testing.T) {
 				},
 			}
 
-			_, err = l.ExecuteScripts(context.Background(), false, script)
+			_, err = l.ExecuteScript(context.Background(), false, script)
 			require.NoError(t, err)
 			assertBalance(t, l, "world", "COIN", core.NewMonetaryInt(0))
 			assertBalance(t, l, "users:001", "COIN", core.NewMonetaryInt(0))
@@ -651,7 +651,7 @@ func TestMonetaryVariableBalance(t *testing.T) {
 				},
 			}
 
-			_, err = l.ExecuteScripts(context.Background(), false, script)
+			_, err = l.ExecuteScript(context.Background(), false, script)
 			require.NoError(t, err)
 			assertBalance(t, l, "B", "USD/2", core.NewMonetaryInt(40))
 			assertBalance(t, l, "D", "USD/2", core.NewMonetaryInt(60))
@@ -689,7 +689,7 @@ func TestMonetaryVariableBalance(t *testing.T) {
  					)`,
 				},
 			}
-			_, err = l.ExecuteScripts(context.Background(), false, script)
+			_, err = l.ExecuteScript(context.Background(), false, script)
 			assert.True(t, ledger.IsScriptErrorWithCode(err, apierrors.ErrInsufficientFund))
 		})
 	})
@@ -722,7 +722,7 @@ func TestMonetaryVariableBalance(t *testing.T) {
 				},
 			}
 
-			_, err = l.ExecuteScripts(context.Background(), false, script)
+			_, err = l.ExecuteScript(context.Background(), false, script)
 			assert.True(t, ledger.IsScriptErrorWithCode(err, ledger.ScriptErrorCompilationFailed))
 			assert.ErrorContains(t, err, "must be non-negative")
 		})
@@ -742,7 +742,7 @@ func TestMonetaryVariableBalance(t *testing.T) {
  					)`,
 				},
 			}
-			_, err := l.ExecuteScripts(context.Background(), false, script)
+			_, err := l.ExecuteScript(context.Background(), false, script)
 			assert.True(t, ledger.IsScriptErrorWithCode(err, apierrors.ErrScriptCompilationFailed))
 		})
 	})
@@ -775,17 +775,16 @@ func TestNewMachineFromScript(t *testing.T) {
 	}
 	_, err := txData.Postings.Validate()
 	require.NoError(t, err)
-	scripts := core.TxsToScriptsData(txData)
-	script := scripts[0].Plain
+	script := txToScriptData(txData)
 
 	h := sha256.New()
-	_, err = h.Write([]byte(script))
+	_, err = h.Write([]byte(script.Plain))
 	require.NoError(t, err)
 	key := h.Sum(nil)
 	keySizeBytes := size.Of(key)
 	require.NotEqual(t, -1, keySizeBytes)
 
-	prog, err := compiler.Compile(script)
+	prog, err := compiler.Compile(script.Plain)
 	require.NoError(t, err)
 	progSizeBytes := size.Of(*prog)
 	require.NotEqual(t, -1, progSizeBytes)
@@ -795,7 +794,7 @@ func TestNewMachineFromScript(t *testing.T) {
 
 		cache := ledger.NewCache(capacityBytes, 1, true)
 
-		m, err := ledger.NewMachineFromScript(script, cache, span)
+		m, err := ledger.NewMachineFromScript(script.Plain, cache, span)
 		require.NoError(t, err)
 		require.NotNil(t, m)
 		cache.Wait()
@@ -803,7 +802,7 @@ func TestNewMachineFromScript(t *testing.T) {
 		require.Equal(t, uint64(1), cache.Metrics.Misses())
 		require.Equal(t, uint64(1), cache.Metrics.KeysAdded())
 
-		m, err = ledger.NewMachineFromScript(script, cache, span)
+		m, err = ledger.NewMachineFromScript(script.Plain, cache, span)
 		require.NoError(t, err)
 		require.NotNil(t, m)
 		cache.Wait()
@@ -817,7 +816,7 @@ func TestNewMachineFromScript(t *testing.T) {
 
 		cache := ledger.NewCache(capacityBytes, 1, true)
 
-		m, err := ledger.NewMachineFromScript(script, cache, span)
+		m, err := ledger.NewMachineFromScript(script.Plain, cache, span)
 		require.NoError(t, err)
 		require.NotNil(t, m)
 		cache.Wait()
@@ -825,7 +824,7 @@ func TestNewMachineFromScript(t *testing.T) {
 		require.Equal(t, uint64(1), cache.Metrics.Misses())
 		require.Equal(t, uint64(0), cache.Metrics.KeysAdded())
 
-		m, err = ledger.NewMachineFromScript(script, cache, span)
+		m, err = ledger.NewMachineFromScript(script.Plain, cache, span)
 		require.NoError(t, err)
 		require.NotNil(t, m)
 		cache.Wait()
@@ -836,6 +835,7 @@ func TestNewMachineFromScript(t *testing.T) {
 }
 
 var execRes []core.ExpandedTransaction
+var execResScript core.ExpandedTransaction
 
 const nbPostings = 1000
 
@@ -850,101 +850,134 @@ func BenchmarkLedger_PostTransactions_Scripts_Single_FixedAccounts(b *testing.B)
 				Amount:      core.NewMonetaryInt(10),
 			})
 		}
+		script := txToScriptData(txData)
 
 		b.ResetTimer()
 
-		res := []core.ExpandedTransaction{}
+		res := core.ExpandedTransaction{}
 
 		for n := 0; n < b.N; n++ {
 			_, err := txData.Postings.Validate()
 			require.NoError(b, err)
-			res, err = l.ExecuteScripts(context.Background(), true, core.TxsToScriptsData(txData)...)
+			res, err = l.ExecuteScript(context.Background(), true, script)
 			require.NoError(b, err)
 			require.Len(b, res, 1)
-			require.Len(b, res[0].Postings, nbPostings)
+			require.Len(b, res.Postings, nbPostings)
 		}
 
-		execRes = res
+		execResScript = res
 		require.Len(b, execRes, 1)
 		require.Len(b, execRes[0].Postings, nbPostings)
 	})
 }
 
-func BenchmarkLedger_PostTransactions_Scripts_Batch_FixedAccounts(b *testing.B) {
-	runOnLedger(func(l *ledger.Ledger) {
-		txsData := newTxsData(1)
-
-		b.ResetTimer()
-
-		res := []core.ExpandedTransaction{}
-
-		for n := 0; n < b.N; n++ {
-			var err error
-			for _, txData := range txsData {
-				_, err := txData.Postings.Validate()
-				require.NoError(b, err)
-			}
-			script := core.TxsToScriptsData(txsData...)
-			res, err = l.ExecuteScripts(context.Background(), true, script...)
-			require.NoError(b, err)
-			require.Len(b, res, 7)
-			require.Len(b, res[0].Postings, 1)
-			require.Len(b, res[1].Postings, 1)
-			require.Len(b, res[2].Postings, 2)
-			require.Len(b, res[3].Postings, 4)
-			require.Len(b, res[4].Postings, 4)
-			require.Len(b, res[5].Postings, 1)
-			require.Len(b, res[6].Postings, 1)
-		}
-
-		execRes = res
-		require.Len(b, execRes, 7)
-		require.Len(b, execRes[0].Postings, 1)
-		require.Len(b, execRes[1].Postings, 1)
-		require.Len(b, execRes[2].Postings, 2)
-		require.Len(b, execRes[3].Postings, 4)
-		require.Len(b, execRes[4].Postings, 4)
-		require.Len(b, execRes[5].Postings, 1)
-		require.Len(b, execRes[6].Postings, 1)
-	})
+type variable struct {
+	name    string
+	jsonVal json.RawMessage
 }
 
-func BenchmarkLedger_PostTransactions_Scripts_Batch_VaryingAccounts(b *testing.B) {
-	runOnLedger(func(l *ledger.Ledger) {
-		b.ResetTimer()
+func txToScriptData(txData core.TransactionData) core.ScriptData {
+	if len(txData.Postings) == 0 {
+		return core.ScriptData{}
+	}
 
-		res := []core.ExpandedTransaction{}
-
-		for n := 0; n < b.N; n++ {
-			b.StopTimer()
-			txsData := newTxsData(n)
-			b.StartTimer()
-			var err error
-			for _, txData := range txsData {
-				_, err := txData.Postings.Validate()
-				require.NoError(b, err)
+	sb := strings.Builder{}
+	monetaryToVars := map[string]variable{}
+	accountsToVars := map[string]variable{}
+	i := 0
+	j := 0
+	for _, p := range txData.Postings {
+		if _, ok := accountsToVars[p.Source]; !ok {
+			if p.Source != core.WORLD {
+				accountsToVars[p.Source] = variable{
+					name:    fmt.Sprintf("va%d", i),
+					jsonVal: json.RawMessage(`"` + p.Source + `"`),
+				}
+				i++
 			}
-			script := core.TxsToScriptsData(txsData...)
-			res, err = l.ExecuteScripts(context.Background(), true, script...)
-			require.NoError(b, err)
-			require.Len(b, res, 7)
-			require.Len(b, res[0].Postings, 1)
-			require.Len(b, res[1].Postings, 1)
-			require.Len(b, res[2].Postings, 2)
-			require.Len(b, res[3].Postings, 4)
-			require.Len(b, res[4].Postings, 4)
-			require.Len(b, res[5].Postings, 1)
-			require.Len(b, res[6].Postings, 1)
 		}
+		if _, ok := accountsToVars[p.Destination]; !ok {
+			if p.Destination != core.WORLD {
+				accountsToVars[p.Destination] = variable{
+					name:    fmt.Sprintf("va%d", i),
+					jsonVal: json.RawMessage(`"` + p.Destination + `"`),
+				}
+				i++
+			}
+		}
+		mon := fmt.Sprintf("[%s %s]", p.Amount.String(), p.Asset)
+		if _, ok := monetaryToVars[mon]; !ok {
+			monetaryToVars[mon] = variable{
+				name: fmt.Sprintf("vm%d", j),
+				jsonVal: json.RawMessage(
+					`{"asset":"` + p.Asset + `","amount":` + p.Amount.String() + `}`),
+			}
+			j++
+		}
+	}
 
-		execRes = res
-		require.Len(b, execRes, 7)
-		require.Len(b, execRes[0].Postings, 1)
-		require.Len(b, execRes[1].Postings, 1)
-		require.Len(b, execRes[2].Postings, 2)
-		require.Len(b, execRes[3].Postings, 4)
-		require.Len(b, execRes[4].Postings, 4)
-		require.Len(b, execRes[5].Postings, 1)
-		require.Len(b, execRes[6].Postings, 1)
-	})
+	sb.WriteString("vars {\n")
+	accVars := make([]string, 0)
+	for _, v := range accountsToVars {
+		accVars = append(accVars, v.name)
+	}
+	sort.Strings(accVars)
+	for _, v := range accVars {
+		sb.WriteString(fmt.Sprintf("\taccount $%s\n", v))
+	}
+	monVars := make([]string, 0)
+	for _, v := range monetaryToVars {
+		monVars = append(monVars, v.name)
+	}
+	sort.Strings(monVars)
+	for _, v := range monVars {
+		sb.WriteString(fmt.Sprintf("\tmonetary $%s\n", v))
+	}
+	sb.WriteString("}\n")
+
+	for _, p := range txData.Postings {
+		m := fmt.Sprintf("[%s %s]", p.Amount.String(), p.Asset)
+		mon, ok := monetaryToVars[m]
+		if !ok {
+			panic(fmt.Sprintf("monetary %s not found", m))
+		}
+		sb.WriteString(fmt.Sprintf("send $%s (\n", mon.name))
+		if p.Source == core.WORLD {
+			sb.WriteString("\tsource = @world\n")
+		} else {
+			src, ok := accountsToVars[p.Source]
+			if !ok {
+				panic(fmt.Sprintf("source %s not found", p.Source))
+			}
+			sb.WriteString(fmt.Sprintf("\tsource = $%s allowing unbounded overdraft\n", src.name))
+		}
+		if p.Destination == core.WORLD {
+			sb.WriteString("\tdestination = @world\n")
+		} else {
+			dest, ok := accountsToVars[p.Destination]
+			if !ok {
+				panic(fmt.Sprintf("destination %s not found", p.Destination))
+			}
+			sb.WriteString(fmt.Sprintf("\tdestination = $%s\n", dest.name))
+		}
+		sb.WriteString(")\n")
+	}
+
+	vars := map[string]json.RawMessage{}
+	for _, v := range accountsToVars {
+		vars[v.name] = v.jsonVal
+	}
+	for _, v := range monetaryToVars {
+		vars[v.name] = v.jsonVal
+	}
+
+	return core.ScriptData{
+		Script: core.Script{
+			Plain: sb.String(),
+			Vars:  vars,
+		},
+		Timestamp: txData.Timestamp,
+		Reference: txData.Reference,
+		Metadata:  txData.Metadata,
+	}
 }
